Added guard to avoid errors if there are no files

+++ clone-image-server/run.sh
@@ -11,11 +11,13 @@ python -m pipenv install
 echo "Removing all running umboxes with images still in this folder."
 INSTANCE_IMAGES="./images/instances/*.qcow2"
 for FILEPATH in ${INSTANCE_IMAGES}; do
-    BASENAME=$(basename $FILEPATH .qcow2)
-    echo "Removing VM with name ${BASENAME}"
-    sudo virsh destroy $BASENAME
-    echo "Removing image file ${FILEPATH}"
-    rm -f $FILEPATH
+    if [ "$FILEPATH" != "$INSTANCE_IMAGES" ]; then
+        BASENAME=$(basename $FILEPATH .qcow2)
+        echo "Removing VM with name ${BASENAME}"
+        sudo virsh destroy $BASENAME
+        echo "Removing image file ${FILEPATH}"
+        rm -f $FILEPATH
+    fi
 done
 
 python -m pipenv run python clone_image_api.py
